// src/utils/ReturnCalcMath/types.ts
export interface ReturnCalcInputs {
  initialInvestment: number;
  recurringInvestment: number;
  recurringFrequency: number;
  timeValue: number;
  timeUnit: number;
  returnRate: number;
  returnRateFrequency: number;
}

export interface ReturnCalcResult {
  finalAmount: number;
  totalInvested: number;
  totalGains: number;
}